requirements.txt
streamlit
pandas
numpy
plotly
streamlit-plotly-events
pillow
